updated icon, reorganized tools, removed description
