+++ sim/app/w/[id]/components/workflow-block/components/sub-block/sub-block.tsx
@@ -202,9 +202,6 @@ export function SubBlock({ blockId, config, isConnecting }: SubBlockProps) {
         </Label>
       )}
       {renderInput()}
-      {config.description && (
-        <p className="text-xs text-muted-foreground mt-1">{config.description}</p>
-      )}
     </div>
   )
 }
