+++ sim/tools/mem0/types.ts
@@ -1,0 +1,9 @@
+import { ToolResponse } from "../types"
+
+export interface Mem0Response extends ToolResponse {
+    output: {
+      ids?: string[]
+      memories?: any[]
+      searchResults?: any[]
+    }
+  }
